chore: extract shared logic into functions
